missing risk measuring code only

+++ app/src/main/java/com/then0wheregirl/santanderproject/model/fund/Year.java
@@ -4,24 +4,24 @@ import com.google.gson.annotations.SerializedName;
 
 public class Year {
 
-    public int fund;
+    public double fund;
 
-    public int getFund() {
+    public double getFund() {
         return fund;
     }
 
-    public void setFund(int fund){
+    public void setFund(double fund){
         this.fund = fund;
     }
 
     @SerializedName("CDI")
-    public int cdi;
+    public double cdi;
 
-    public int getCdi(){
+    public double getCdi(){
         return cdi;
     }
 
-    public void setCdi(int cdi){
+    public void setCdi(double cdi){
         this.cdi = cdi;
     }
 }
